+++ keras_nlp/models/gpt2/gpt2_causal_lm_preprocessor.py
@@ -1,0 +1,34 @@
+# Copyright 2022 The KerasNLP Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+import tensorflow as tf
+
+from keras_nlp.models.gpt2.gpt2_preprocessor import GPT2Preprocessor
+from keras_nlp.utils.keras_utils import pack_x_y_sample_weight
+
+
+class GPT2CausalLMPreprocessor(GPT2Preprocessor):
+    def call(self, x, y=None, sample_weight=None):
+        token_ids = self.tokenizer(x)
+        mask = tf.ones_like(token_ids, dtype=tf.bool)
+        mask = mask.to_tensor(shape=(None, self.sequence_length))
+        token_ids = token_ids.to_tensor(shape=(None, self.sequence_length))
+        x = {
+            "token_ids": token_ids[:, :-1],
+            "padding_mask": mask[:, 1:],
+        }
+
+        y = token_ids[:, 1:]
+        sample_weight = mask[:, 1:]
+
+        return pack_x_y_sample_weight(x, y, sample_weight)
